chore: refine the mod.rs
