+++ src/query/service/src/pipelines/processors/transforms/mod.rs
@@ -17,12 +17,6 @@ mod chunk_operator;
 pub(crate) mod hash_join;
 mod transform_addon;
 mod transform_aggregator;
-use common_pipeline_transforms::processors::transforms::transform;
-use common_pipeline_transforms::processors::transforms::transform_block_compact;
-use common_pipeline_transforms::processors::transforms::transform_compact;
-use common_pipeline_transforms::processors::transforms::transform_expression;
-use common_pipeline_transforms::processors::transforms::transform_sort_merge;
-use common_pipeline_transforms::processors::transforms::transform_sort_partial;
 mod transform_cast_schema;
 mod transform_create_sets;
 mod transform_dummy;
@@ -39,6 +33,12 @@ pub use aggregator::AggregatorParams;
 pub use aggregator::AggregatorTransformParams;
 pub use chunk_operator::ChunkOperator;
 pub use chunk_operator::CompoundChunkOperator;
+use common_pipeline_transforms::processors::transforms::transform;
+use common_pipeline_transforms::processors::transforms::transform_block_compact;
+use common_pipeline_transforms::processors::transforms::transform_compact;
+use common_pipeline_transforms::processors::transforms::transform_expression;
+use common_pipeline_transforms::processors::transforms::transform_sort_merge;
+use common_pipeline_transforms::processors::transforms::transform_sort_partial;
 pub use common_pipeline_transforms::processors::ExpressionExecutor;
 pub use hash_join::HashJoinDesc;
 pub use hash_join::HashJoinState;
