+++ src/query/service/src/sql/planner/optimizer/rule/mod.rs
@@ -12,15 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::fmt::Display;
-use std::fmt::Formatter;
-
-use common_exception::Result;
-
-use crate::sql::optimizer::SExpr;
-
 mod factory;
 mod rewrite;
+#[allow(clippy::module_inception)]
+mod rule;
 mod rule_implement_get;
 mod rule_implement_hash_join;
 mod rule_set;
@@ -28,74 +23,9 @@ mod transform;
 mod transform_state;
 
 pub use factory::RuleFactory;
+pub use rule::Rule;
+pub use rule::RuleID;
+pub use rule::RulePtr;
 pub use rule_set::AppliedRules;
 pub use rule_set::RuleSet;
 pub use transform_state::TransformState;
-
-pub type RulePtr = Box<dyn Rule>;
-
-pub trait Rule {
-    fn id(&self) -> RuleID;
-
-    fn apply(&self, s_expr: &SExpr, state: &mut TransformState) -> Result<()>;
-
-    fn pattern(&self) -> &SExpr;
-}
-
-#[derive(Debug, Copy, Clone, Eq, PartialEq, Hash)]
-pub enum RuleID {
-    // Rewrite rules
-    NormalizeScalarFilter,
-    NormalizeDisjunctiveFilter,
-    PushDownFilterEvalScalar,
-    PushDownFilterJoin,
-    PushDownFilterScan,
-    PushDownLimitOuterJoin,
-    PushDownLimitSort,
-    PushDownLimitScan,
-    PushDownSortScan,
-    EliminateEvalScalar,
-    EliminateFilter,
-    MergeEvalScalar,
-    MergeFilter,
-    SplitAggregate,
-    FoldCountAggregate,
-
-    // Exploration rules
-    CommuteJoin,
-    LeftAssociateJoin,
-    RightAssociateJoin,
-
-    // Implementation rules
-    ImplementGet,
-    ImplementHashJoin,
-}
-
-impl Display for RuleID {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        match self {
-            RuleID::PushDownFilterEvalScalar => write!(f, "PushDownFilterEvalScalar"),
-            RuleID::PushDownFilterJoin => write!(f, "PushDownFilterJoin"),
-            RuleID::PushDownFilterScan => write!(f, "PushDownFilterScan"),
-            RuleID::PushDownLimitOuterJoin => write!(f, "PushDownLimitOuterJoin"),
-            RuleID::PushDownLimitSort => write!(f, "PushDownLimitSort"),
-            RuleID::PushDownLimitScan => write!(f, "PushDownLimitScan"),
-            RuleID::PushDownSortScan => write!(f, "PushDownSortScan"),
-            RuleID::EliminateEvalScalar => write!(f, "EliminateEvalScalar"),
-            RuleID::EliminateFilter => write!(f, "EliminateFilter"),
-            RuleID::MergeEvalScalar => write!(f, "MergeEvalScalar"),
-            RuleID::MergeFilter => write!(f, "MergeFilter"),
-            RuleID::NormalizeScalarFilter => write!(f, "NormalizeScalarFilter"),
-            RuleID::SplitAggregate => write!(f, "SplitAggregate"),
-            RuleID::NormalizeDisjunctiveFilter => write!(f, "NormalizeDisjunctiveFilter"),
-            RuleID::FoldCountAggregate => write!(f, "FoldCountAggregate"),
-
-            RuleID::CommuteJoin => write!(f, "CommuteJoin"),
-            RuleID::LeftAssociateJoin => write!(f, "LeftAssociateJoin"),
-            RuleID::RightAssociateJoin => write!(f, "RightAssociateJoin"),
-
-            RuleID::ImplementGet => write!(f, "ImplementGet"),
-            RuleID::ImplementHashJoin => write!(f, "ImplementHashJoin"),
-        }
-    }
-}
